Backend/requirements.txt
openai-whisper
flask
flask-cors
gunicorn
